+++ components/frontend/src/context/Permissions.test.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { ReadOnlyContext, ReadOnlyOrEditable } from './ReadOnly';
+import { Permissions, ReadOnlyOrEditable } from './Permissions';
 
 function MockComponent() { return ("Hi") }
 function MockComponent2() { return ("Two") }
 
-describe("<ReadOnly />", () => {
-  it('provides the read only context to components', () => {
+describe("<Permissions />", () => {
+  it('provides the permissions context to components', () => {
     const wrapper = shallow(
-      <ReadOnlyContext.Provider value={['mockPermission']}>
-        <ReadOnlyContext.Consumer>{(permissions) => (<MockComponent permissions={permissions} />)}</ReadOnlyContext.Consumer>
-      </ReadOnlyContext.Provider>);
+      <Permissions.Provider value={['mockPermission']}>
+        <Permissions.Consumer>{(permissions) => (<MockComponent permissions={permissions} />)}</Permissions.Consumer>
+      </Permissions.Provider>);
     expect(wrapper.dive().find('MockComponent').prop("permissions")).toStrictEqual(['mockPermission']);
   });
 });
@@ -18,26 +18,26 @@ describe("<ReadOnly />", () => {
 describe("<ReadOnlyOrEditable />", () => {
   it('shows the read only component if no permissions are present', () => {
     const wrapper = shallow(
-      <ReadOnlyContext.Provider value={[]}>
+      <Permissions.Provider value={[]}>
         <ReadOnlyOrEditable requiredPermissions={['mockPermission']} readOnlyComponent={<MockComponent/>} editableComponent={<MockComponent2/>} />
-      </ReadOnlyContext.Provider>)
+      </Permissions.Provider>)
     expect(wrapper.dive().dive().name()).toBe("MockComponent");
   });
   it('shows the read only component if not all permissions are present', () => {
     const wrapper = shallow(
-      <ReadOnlyContext.Provider value={['mockPermission']}>
+      <Permissions.Provider value={['mockPermission']}>
         <ReadOnlyOrEditable 
           requiredPermissions={['mockPermission', 'mockPermission1']}
           readOnlyComponent={<MockComponent/>}
           editableComponent={<MockComponent2/>} />
-      </ReadOnlyContext.Provider>)
+      </Permissions.Provider>)
     expect(wrapper.dive().dive().name()).toBe("MockComponent");
   });
   it('shows the editable only component', () => {
     const wrapper = shallow(
-      <ReadOnlyContext.Provider value={['mockPermission']}>
+      <Permissions.Provider value={['mockPermission']}>
         <ReadOnlyOrEditable requiredPermissions={['mockPermission']} readOnlyComponent={<MockComponent/>} editableComponent={<MockComponent2/>} />
-      </ReadOnlyContext.Provider>);
+      </Permissions.Provider>);
     expect(wrapper.dive().dive().name()).toBe("MockComponent2");
   });
 });