+++ components/frontend/src/fields/DateInput.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Form } from 'semantic-ui-react';
 import { DateInput as CalendarDateInput } from 'semantic-ui-calendar-react';
-import { ReadOnlyOrEditable } from '../context/ReadOnly';
+import { ReadOnlyOrEditable } from '../context/Permissions';
 import { Input } from './Input';
 
 function EditableDateInput(props) {