+++ components/frontend/src/report/Report.js
@@ -4,7 +4,7 @@ import { Subjects } from '../subject/Subjects';
 import { Tag } from '../widgets/Tag';
 import { MetricSummaryCard } from '../dashboard/MetricSummaryCard';
 import { CardDashboard } from '../dashboard/CardDashboard';
-import { ReadOnlyContext } from '../context/ReadOnly';
+import { accessGranted, EDIT_REPORT_PERMISSION, Permissions } from '../context/Permissions';
 import { set_report_attribute } from '../api/report';
 import { get_subject_name, useURLSearchQuery } from '../utils';
 import { ReportTitle } from './ReportTitle';
@@ -31,13 +31,13 @@ function ReportDashboard(props) {
         );
     }
     return (
-        <ReadOnlyContext.Consumer>{(readOnly) => (
+        <Permissions.Consumer>{(permissions) => (
             <CardDashboard
                 cards={subject_cards().concat(tag_cards())}
                 initial_layout={props.report.layout || []}
-                save_layout={function (layout) { if (!readOnly) { set_report_attribute(props.report.report_uuid, "layout", layout, props.reload) } }}
+                save_layout={function (layout) { if (accessGranted(permissions, [EDIT_REPORT_PERMISSION])) { set_report_attribute(props.report.report_uuid, "layout", layout, props.reload) } }}
             />)}
-        </ReadOnlyContext.Consumer>
+        </Permissions.Consumer>
     )
 }