+++ components/frontend/src/source/SourceEntityDetails.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import { EDIT_ENTITY_PERMISSION, ReadOnlyContext } from '../context/ReadOnly';
+import { EDIT_ENTITY_PERMISSION, Permissions } from '../context/Permissions';
 import * as source from '../api/source';
 import { SourceEntityDetails } from './SourceEntityDetails';
 
@@ -10,17 +10,17 @@ describe('<SourceEntityDetails />', () => {
     beforeEach(() => { source.set_source_entity_attribute = jest.fn(); });
     it('invokes the callback on changing the status', () => {
         const wrapper = mount(
-            <ReadOnlyContext.Provider value={[EDIT_ENTITY_PERMISSION]}>
+            <Permissions.Provider value={[EDIT_ENTITY_PERMISSION]}>
                 <SourceEntityDetails entity={{key: "key"}} status="unconfirmed" name="violation" />
-            </ReadOnlyContext.Provider>);
+            </Permissions.Provider>);
         wrapper.find("div.item").at(1).simulate("click");
         expect(source.set_source_entity_attribute).toHaveBeenCalled();
     });
     it('invokes the callback on changing the comment', () => {
         const wrapper = mount(
-            <ReadOnlyContext.Provider value={[EDIT_ENTITY_PERMISSION]}>
+            <Permissions.Provider value={[EDIT_ENTITY_PERMISSION]}>
                 <SourceEntityDetails entity={{key: "key"}} status="unconfirmed" name="violation" />
-            </ReadOnlyContext.Provider>);
+            </Permissions.Provider>);
         wrapper.find("textarea").at(0).simulate("change", {target: {value: "comment"}});
         wrapper.find("TextInput").at(0).simulate("submit");
         expect(source.set_source_entity_attribute).toHaveBeenCalled();