+++ components/frontend/src/subject/SubjectFooter.js
@@ -1,6 +1,6 @@
 import { Table } from "semantic-ui-react";
 import { add_metric, copy_metric, move_metric } from "../api/metric";
-import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from "../context/ReadOnly";
+import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from "../context/Permissions";
 import { AddButton, CopyButton, MoveButton } from "../widgets/Button";
 import { metric_options } from "../widgets/menu_options";