+++ components/frontend/src/subject/Subjects.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { EDIT_REPORT_PERMISSION, ReadOnlyContext } from '../context/ReadOnly';
+import { EDIT_REPORT_PERMISSION, Permissions } from '../context/Permissions';
 import { Subjects } from './Subjects';
 
 const datamodel = { subjects: { subject_type: { name: "Subject type" } }, metrics: { metric_type: { tags: [] } } }
@@ -33,9 +33,9 @@ describe("<Subjects />", () => {
         expect(wrapper.find("Subject").length).toBe(1);
     });
     it('shows the add subject button when editable', () => {
-        const wrapper = shallow(<ReadOnlyContext.Provider value={[EDIT_REPORT_PERMISSION]}>
+        const wrapper = shallow(<Permissions.Provider value={[EDIT_REPORT_PERMISSION]}>
             <Subjects datamodel={datamodel} reports={[report]} report={report} tags={[]} history={mockHistory} hidden_columns={[]} hide_metrics_not_requiring_action={false} />
-        </ReadOnlyContext.Provider>);
+        </Permissions.Provider>);
         expect(wrapper.find("Subjects").dive().find("ReadOnlyOrEditable").dive().find("ContextConsumer").dive().find("AddButton").prop("item_type")).toStrictEqual("subject");
     });
     it('hides metrics not requiring action', () => {