+++ components/frontend/src/utils.test.js
@@ -1,6 +1,6 @@
 import * as react_semantic_toast from 'react-semantic-toasts';
 import { getUserPermissions, get_metric_tags, get_metric_target, get_source_name, get_subject_name, nice_number, scaled_number, show_message, format_minutes } from './utils';
-import { EDIT_REPORT_PERMISSION, EDIT_ENTITY_PERMISSION } from './context/ReadOnly';
+import { EDIT_REPORT_PERMISSION, EDIT_ENTITY_PERMISSION } from './context/Permissions';
 
 jest.mock("react-semantic-toasts");